requirements.txt
kabbes_client
kabbes_config
kabbes_pypi_builder
kabbes_py_starter>=0.8.0
kabbes_parent_class>=0.5.0
kabbes_dir_ops>=0.7.0
kabbes_repository_generator>=0.8.0
kabbes_pypi_builder>=0.8.0
sphinx
sphinx-autodoc-typehints
sphinx_rtd_theme